chore: organizando arquivos
