+++ exercicios/para-casa/anna-berto/Bank.js
@@ -1,0 +1,19 @@
+class Bank {
+  bankCode;
+  bankName;
+  qtdPersons;
+  #transferRate;
+  qtdPersons = 0;
+  constructor(bankCode, bankName, transferRate) {
+    this.bankCode = bankCode;
+    this.bankName = bankName;
+    this.#transferRate = transferRate;
+    this.constructor.bankCreated.push({
+      codBankCreated: this.bankCode,
+      qtdPersons: this.qtdPersons,
+    });
+  }
+  static bankCreated = [];
+}
+
+module.exports = { Bank };
