+++ exercicios/para-casa/anna-berto/Person.js
@@ -1,0 +1,9 @@
+class Person {
+  name;
+  cpf;
+  constructor(name, cpf) {
+    this.name = name;
+    this.cpf = cpf;
+  }
+};
+module.exports = { Person };
